Split out swh.alter.cli.remove() in a new module
The removal operation is at the core of `swh-alter`. It was currently
only implemented in the code managing the CLI interface. This is far
from optimal as we may create other frontends, and it also intertwined
code to display progress and confirm with actual operations. Before
adding more code to perform the actual removing this was worth a
refactoring to avoid having the function grow out of hands.

A new module `swh.alter.operations` now has an helper class `Remover`
with methods for the various steps of the removal operation. It uses
`logging` facilities for progress, we can later be formatted nicely
be the `cli` module.

The tests have been rearranged so test of actual operations are done in
`test_operations.py`. For removals, `test_cli.py` will mock parts of the
`Remover` class to ensure that the right arguments are passed depending
on the command-line options.