Remove objects that we know are in the recovery bundle
Instead of passing the list of SWHIDs to
`Remover.create_recovery_bundle` and then pass the same list to
`Remover.remove`, the `Remover` object will hold the list of objects to
remove. This avoid potential mismatches.

This is implemented by adding a registration callback in
`RecoveryBundleCreator`. This paves the way for removal from the
journal in which objects are not all referenced using SWHIDs.