Add support for removal from Kafka (aka the journal)
When creating the recovery bundle, we not only record SWHIDs of objects
to remove but also their `unique_key()`, sorted by object type. We can
then easily call `KafkaJournalWriter.delete()` to remove entries from
the journal.

The registration callback is called for all objects written to the
bundle, even those which do not have a SWHID.