Display an error when the graph is unavailable
We really need to have the graph available to look up which candidates
can be removed or not, as the `object_references` tables only contain data
since the last export. So we display an error when the graph is not
available at the beginning of the remove command.

The initial connection on RemoteGraphClient instantiation was introduced
in swh-graph 3.1.0. Bump the dependency accordingly.

Thanks to vlorentz for the review and suggestion to use ClickException.

Closes #13