Instantiate SecretSharing early in `remove` command
When trying to do a removal with a wrong secret sharing config,
we used to do all the work of figuring out which objects can be
removed, and then crash, without a way to resume. Instead, we now
instantiate `SecretSharing` early from the configuration. We also
take the opportunity to display en error message instead of a full stack
trace.

Closes: #16