Configure each facility according to their use
In order to remove potential sources of confusion, we now ask users
to have different entries for the configuration for the facilities that
will be used to read, restore and remove objects.

- `storage` and `graph` will be used for reading.
- `restoration_storage` will be used for restoration.
- `removal_storages` and `removal_journals` will be used for removals.

This means they now all have to be defined in order for `swh alter
remove` to work properly. The upside is that the required permissions
for each facility are more clearly defined.