Fix storage configuration in tests
The `restoration_storage` configuration introduced in 29bce6a34a2
in the `remove_config` used to tests removal was wrong. This does not
seem to be a problem in practice as we define a specific configuration
when testing restoration, but better safe than sorry.

There is no need to specify that `objstorage` uses the `memory` class
as it is implied when using the `memory` class for storage. But in case
we need to do a restore, the `journal_writer` needs to be functional, so
we need to use the `memory` class as well there.