Refactor logging in CLI and in recovery_bundle module
As we are working our way to enable resuming removals from recovery
bundles, we will need to be able to perform a restore in case
something bad happens during the removal, similarly to what happens in
`swh alter remove`.

We want to avoid repeating the code logging the results of the restore
operation. So instead of having it lying the `cli` module, it has now been
moved to the `recovery_bundle` module. This means we have to setup the
logger to properly display the output on the console when the operation
is run from the CLI.

Such setup in the beginning of `cli.remove()`, but there is actually no
reason to not benefit from it when doing other operations from the command-line,
like when doing a restore so we can display the same statistics.
We move the setup to the `swh alter` command line group initialization.

In return, we have to adapt several of our tests to call `alter_cli_group` instead
of the various function. We also add a new fixture that should go
alongside, `capture_output`, that isolates our tests by using dedicated
loggers.