cli: Stop assuming a default configuration
Having a default configuration to run our command-line interface
does not really make sense. Now that we have more features, the
configuration has to be specific to the environment we run in.

Instead, stop displaying a stack trace when the environment variable
`SWH_CONFIG_FILENAME` is not defined and just print an error message
instead.