restore: flush storage at every loop
In the production-like setting, we've deployed swh-alter with a buffered
storage for the restore process. If the buffer storage is not
flushed (at least on exit), it just loses the added objects.

To honor the swh-alter chunking mechanism, we call storage.flush() after
every call to an add function. We make sure to run the recovery_bundle
tests with a buffered storage to trigger the same code paths.