Error out when the inventory phase gets stuck
While it is unlikely, we could have missing objects in the storage that
makes us unable to complete the inventory phase. This can happen in case
of lagging replayers for example.

Instead of getting in an infinite loop, we now raise a dedicated
exception, turned into an error message, if that happens.