Use a storage config closer to production in tests
On production, `swh-alter` will be configured to run against a `storage`
that has a populated `object_references` table and a
`restoration_storage` that has a buffering proxy configured. Let’s run
all our tests using such configuration in order to be closer to what
the behaviors that should happen during real usage.