Fix spurious warnings when an object cannot be found in an objstorage
We should not get an error message when an object cannot be found in a
single objstorage.  More objstorages have been added as time went on.
But while old objects have not been moved to newer objstorages, they are
still available as content is retrieved by querying multiple objstorages
in succession.

The real problem is when not a single objstorage has a particular object.
We now only print a message in this case.