Add `--known-missing` and `--known-missing-file` options
Once in a while, objects are missing from the storage. It should not
happen but it does: previous removals, replayer errors, corruptions…

`swh alter remove` and `swh alter recovery-bundle resume-removal` now
accepts the new options `--known-missing` and `--known-missing-file` to
handle missing objects.

`--known-missing` is followed by a SWHID and can appear multiple times.
`--known-missing-file` is followed by the path (or `-` for stdin) of a
file with one SWHID per line (skipping empty lines or lines starting
with `#`).

Known missing objects will be skipped during the “inventory” phase and
mark as unremovable during the “removable” phase. Due to the latter,
they will not be used when creating the recovery bundle, and therefore
skipped from removal as well.