Add `requested` and `referencing` fields to recovery bundle
Experiments have shown that it would be beneficial to add two fields
to the recovery bundle manifests:

- `requested`: list the origin URLs or SWHIDs which are at the root of
  the removal. Basically what is given to `swh alter remove`.

- `referencing`: list of SWHIDs referenced by objects in the bundle that
  have not been removed from the archive.

The latter helps us figure out the order in which recovery bundles
should be recovered or deleted (e.g. after expiration). For example,
picture the following scenario with repositories A and B which is a fork
of A:

1. Remove project A. Common objects with B will not be removed.
2. Remove project B. All objects have been removed.
3. Restore bundle of project A. **Oops**, we now have objects
   referencing other objects that are only in the recovery bundle for B.

At least, we want to have a warning before the bundle for A is restored.

After these changes, the manifest is at version 3.

Implementation-wise, `referencing` is populated using the
RemovableSubgraph by looking for unremovable successors to objects found
removable.