Use a dedicated object for Remover.get_removable() results
In a latter commit, we will want to return multiple values from
`Remover.get_removable()` which we will want to use in
`Remover.create_recovery_bundle()`. In order to better show that
the result of one method must be passed to the other, let’s use
a dedicated object `Removable` (using `NamedTuple` for simplicity).