Ask for confirmation when restoring a bundle referencing missing objects
Recovery bundle manifests now contain a list of objects outside the
bundle that are referenced by objects in the bundle. We make use of
this list before restoring a bundle to see if any of these objects are
missing from the storage. In this is the case, a confirmation is
required before proceeding as it will create “dangling pointers” in the
Merkle DAG.