Error out when an origin can’t be found
When using `swh alter remove` or `swh alter list-candidates`,
we used to only display a warning when an origin cannot be found in the
storage. But in both cases, it is more likely to be an error on the
command-line that anything else. So instead of offering to pursue
the removal operation, let’s display an error and exit instead.