Ensure we end the inventory progressbar on 0 left
We forgot that `Lister._iter_inventory_candidates()` would end the
iteration when 0 objects remain to be looked up. This would result in
moving to the removable step while still showing a progressbar
displaying “4 left to look up”.

We now always update the progress bar saying that we are done after
leaving the loop in `Lister.inventory_candidates()`.