Break down the displayed number of objects to remove by type
Instead of asking “Proceed with removing 98 SWHIDs?” which turns out to
be hard to interpret, we now display the number of objects to remove for
each type having a SWHID. We also mention that some objects do not have
a SWHID (as they will add to the counts shown later).

We also now prefer to display these numbers rather than a list of SWHID
when `--dry-run=stop-before-recovery-bundle` is specified. Getting the
list of SWHIDs can be done using `swh alter list-candidates`.