Handle missing data in objstorage
When a Content was missing from the objstorage – because it had been
manually erased before or in a case of corruption – we used to crash
with a ValueError. Instead, we now display a proper error message.
The message suggests using the `--allow-empty-content-objects` flag.
When specified, getting no data will display a warning and a Content
object with no data will be recorded in the recovery bundle.

Closes #22