Add a --no-recompute flag to the 'handle-removal-notification remove' command
This allows to apply a removal request received via the notification
system without recomputing locally the set of SWHIDs to remove, using
the list sent along with the notification message. It allows a mirror to
apply the removal notification exactly as it has been applied on the
main archive, and more importantly, to apply it without the need of an
accessible swh-graph instance.